backend/requirements.txt
astroid==2.2.5
defusedxml==0.6.0
Django==2.2.1
django-filter==2.1.0
django-rest-registration==0.4.4
djangorestframework==3.9.4
djangorestframework-xml==1.4.0
greenlet==0.4.15
isort==4.3.18
lazy-object-proxy==1.4.0
mccabe==0.6.1
msgpack==0.6.1
pylint==2.3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
wrapt==1.11.1
